requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
facenet-pytorch>=2.5.3
numpy>=1.24.0
Pillow>=9.5.0